make full use of the mini lcd v3 except sd card
including beeper/buzzer, lcd, rgb....except for SD card.

+++ bluepill_makerbase_miniLCD12864_v3/Core/Inc/stm32_u8g2_minilcd12864_v3.h
@@ -68,7 +68,7 @@ public:
     SPI_HandleTypeDef * get_spi();
 
 
-    void draw_test_screen( uint16_t fps  ) ;
+    void draw_test_screen( uint8_t fps  ) ;
     void  setup_u8g2_stm32(   ) ;
 
 };
